+++ src/lib/project/migration/versions/410to420.ts
@@ -34,3 +34,32 @@ export const migrate410r1To420r1: ProjectResourcesMigration = {
   to: { version: "4.2.0", release: "1" },
   migrationFn: createScriptEventsMigrator(migrateFrom410r1To420r1Event),
 };
+
+export const migrateFrom420r1To420r2Event: ScriptEventMigrationFn = (
+  scriptEvent
+) => {
+  if (scriptEvent.args && scriptEvent.command === "EVENT_WAIT") {
+    const args: Record<string, unknown> = { ...scriptEvent.args };
+    // Convert to constvalue
+    args["frames"] = {
+      type: "number",
+      value: typeof args["frames"] === "number" ? args["frames"] : 30,
+    };
+    args["time"] = {
+      type: "number",
+      value: typeof args["time"] === "number" ? args["time"] : 0.5,
+    };
+
+    return {
+      ...scriptEvent,
+      args,
+    };
+  }
+  return scriptEvent;
+};
+
+export const migrate420r1To420r2: ProjectResourcesMigration = {
+  from: { version: "4.2.0", release: "1" },
+  to: { version: "4.2.0", release: "2" },
+  migrationFn: createScriptEventsMigrator(migrateFrom420r1To420r2Event),
+};