+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved collision handling for ladders and one-way platforms in the Platformer scene type. Ladders now use the player's bottom edge for anchoring, and one-way platforms no longer snap the player to the platform when colliding from below [@Steinbeuge](https://github.com/Steinbeuge)
 - Ladder collision tile now only visible on Platform scenes by default, edit `engine.json` to add per scene collision tile types
 - New instances of prefabs use prefab's name by default
+- Update "Wait" event to support using variable values for wait time [@pau-tomas](https://github.com/pau-tomas))
 - Updated Polish localisation. [@ReptiIe](https://github.com/ReptiIe)
 - Updated Japanese localisation. [@tomo666](https://github.com/tomo666)